requirements.txt
annotated-types==0.7.0
asgiref==3.9.1
black==25.1.0
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
contextlib2==21.6.0
cryptography==45.0.5
Django==5.2.4
django-jazzmin==3.0.1
django-ninja==1.4.3
django-ninja-extra==0.30.1
django-ninja-jwt==5.3.7
django-phonenumber-field==8.1.0
hubspot-api-client==12.0.0
idna==3.10
injector==0.22.0
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
phonenumberslite==9.0.10
platformdirs==4.3.8
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
requests==2.32.4
six==1.17.0
sqlparse==0.5.3
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
